caddyfile/proxy backend cleanup (#391)
