+++ etc/Caddyfile
@@ -1,4 +1,4 @@
 http:// {
-	bind zrok/{{ .Token }}
+	bind {{ .ZrokBindAddress }}
 	reverse_proxy 127.0.0.1:3000
 }
